create_db only deletes old db if it already exists

+++ backend/create_db.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python3
 
-import json
 import os
 import sqlite3
 
 from update_db_image_url import update_db_image_url
 
-# remove old db
-os.remove("filaments.db")
+# remove old db if exists
+if os.path.exists("filaments.db"):
+    os.remove("filaments.db")
 
 # creating db
 conn = sqlite3.connect("filaments.db")
